Storage rework WIP

+++ src/ifcparse/IfcException.h
@@ -57,7 +57,7 @@ class IFC_PARSE_API IfcAttributeOutOfRangeException : public IfcException {
 class IFC_PARSE_API IfcInvalidTokenException : public IfcException {
   public:
     IfcInvalidTokenException(
-        int token_start,
+        size_t token_start,
         const std::string& token_string,
         const std::string& expected_type)
         : IfcException(
@@ -65,7 +65,7 @@ class IFC_PARSE_API IfcInvalidTokenException : public IfcException {
               boost::lexical_cast<std::string>(token_start) +
               " invalid " + expected_type) {}
     IfcInvalidTokenException(
-        int token_start,
+        size_t token_start,
         char character)
         : IfcException(
               std::string("Unexpected '") + std::string(1, character) + "' at offset " +
